Scoped enums.

+++ common/lc_context.cpp
@@ -684,7 +684,7 @@ void lcContext::SetColorIndex(int ColorIndex)
 
 void lcContext::SetColorIndexTinted(int ColorIndex, lcInterfaceColor InterfaceColor, float Weight)
 {
-	const lcVector3 Color(gColorList[ColorIndex].Value * Weight + gInterfaceColors[InterfaceColor] * (1.0f - Weight));
+	const lcVector3 Color(gColorList[ColorIndex].Value * Weight + gInterfaceColors[static_cast<int>(InterfaceColor)] * (1.0f - Weight));
 	SetColor(lcVector4(Color, gColorList[ColorIndex].Value.w));
 }
 
@@ -705,7 +705,7 @@ void lcContext::SetEdgeColorIndexTinted(int ColorIndex, const lcVector4& Tint)
 
 void lcContext::SetInterfaceColor(lcInterfaceColor InterfaceColor)
 {
-	SetColor(gInterfaceColors[InterfaceColor]);
+	SetColor(gInterfaceColors[static_cast<int>(InterfaceColor)]);
 }
 
 lcVertexBuffer lcContext::CreateVertexBuffer(int Size, const void* Data)
